refactor(Menu): move v1/v2 to src
