+++ packages/components/src/Menu/_docs/examples.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { action } from '@storybook/addon-actions'
 import { Icon } from '~components/__rc__/Icon'
-import { MenuItem, MenuList } from '../'
+import { MenuItem, MenuList } from '..'
 
 export const exampleActionButtonPropsButton = {
   label: 'Edit Survey',
