+++ packages/components/src/Menu/subcomponents/StatelessMenu/StatelessMenu.module.scss
