+++ packages/components/src/TitleBlockZen/subcomponents/MobileActions.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useEffect, useState } from 'react'
 import classnames from 'classnames'
 import { FocusOn } from 'react-focus-on'
 import type { ButtonProps } from '~components/Button'
-import { MenuItem, MenuList, MenuHeading } from '~components/__actions__/v2'
+import { MenuItem, MenuList, MenuHeading } from '~components/Menu'
 import { Icon } from '~components/__rc__/Icon'
 import { TITLE_BLOCK_ZEN_OTHER_ACTIONS_HTML_ID } from '../constants'
 import {
