+++ packages/components/src/TitleBlockZen/subcomponents/Toolbar.tsx
@@ -5,7 +5,7 @@
 import React from 'react'
 import classnames from 'classnames'
 import type { ButtonProps } from '~components/Button'
-import { MenuProps } from '~components/__actions__/v2'
+import { MenuProps } from '~components/Menu'
 import styles from './Toolbar.module.scss'
 
 type ToolbarProps = {
