+++ packages/components/src/TitleBlockZen/types.ts
@@ -1,7 +1,7 @@
 import { CompanyAvatarProps, GenericAvatarProps } from '~components/Avatar'
 import type { ButtonProps, CustomButtonProps } from '~components/Button'
+import { MenuItemProps } from '~components/Menu'
 import { Select } from '~components/Select'
-import { MenuItemProps } from '~components/__actions__/v2'
 import { NavigationTabProps } from './subcomponents/NavigationTabs'
 
 /**
