--- packages/components/src/__actions__/Menu/v2/index.ts
@@ -1,6 +1,0 @@
-export * from '../v1/Menu'
-export * from '../v1/subcomponents/MenuList'
-export * from '../v1/subcomponents/MenuDropdown'
-export * from '../v1/subcomponents/MenuItem'
-export * from '../v1/subcomponents/MenuHeading'
-export * from '../v1/subcomponents/StatelessMenu'
