--- packages/components/src/__actions__/v1.ts
@@ -1,1 +1,0 @@
-export * from './Menu/v1'
