--- packages/components/src/__actions__/v2.ts
@@ -1,1 +1,0 @@
-export * from './Menu/v2'
