+++ packages/components/src/index.ts
@@ -31,6 +31,7 @@ export * from './Label'
 export * from './LabelledMessage'
 export * from './LikertScaleLegacy'
 export * from './Loading'
+export * from './Menu'
 export * from './Modal'
 export * from './MultiSelect'
 export * from './Notification'
@@ -60,5 +61,5 @@ export * from './utils'
 export * from './VisuallyHidden'
 export * from './Well'
 export * from './Workflow'
-export * from './__actions__/Menu/v1'
+export * from './Menu'
 export * from './Button'
