+++ packages/components/src/v1-actions.ts
@@ -1,1 +1,2 @@
 export * from './Button'
+export * from './Menu'
